move connection fixtures into the functional scope
